analyzer/requirements.txt
marshmallow>=3.0
kombu==4.6.3
nameko>=2.11
python-box>=3.4
requests>=2.22
